Move components to a components folder in frontend service
