+++ frontend/be-compliant/src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import { MainTableComponent } from "./Table";
+import { MainTableComponent } from "./components/table/Table";
 
 
 function App() {
